Create English Translation Conclussion
The Smart Contract within the scope, used to interact with the token
VSION of CRIPTOVISION, has gone through a security audit process
through manual procedures in conjunction with tools
automated.


Based on the tests carried out throughout this audit process, it is
can conclude that the security status of the Smart Contract is HIGH, being
the level of risk to which it is subjected is LOW.


It is recommended to resolve the low-impact vulnerabilities detected and,
as far as possible, correct the problems of an informative nature of
for the optimization of the contract.


It should be noted that the code object of this audit has not been
fully commented. Although comments are appreciated in the part of the
contract referring to libraries and standard interfaces, it is recommended that
all the functions, especially the public ones, are commented, as
recommended in the solidity programming language style guide. This
This way, the code is intended to be easier to read and understand by
from other developers, in order to add new features or correct
failures.


We remember that once the contract is displayed in an address
specific cannot be modified or deleted, which causes the
presence of vulnerabilities must be corrected prior to deployment.
Therefore, it is recommended to have exhaustive unit tests. The utilization
of this type of test could prevent unexpected failures from occurring in the
contracts, ensuring that they are deployed with maximum security and
possible quality.